requirements.txt
asgiref
certifi
charset-normalizer
click
detect
Django
docopt
filelock
fsspec
huggingface-hub
idna
image
Jinja2
joblib
langdetect
MarkupSafe
mpmath
networkx
nltk
num2words
numpy
opensearch
packaging
Pillow
PorterStemmer
PyYAML
regex
requests
safetensors
scikit-learn
scipy
six
sqlparse
stopwords
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
sentence-transformers
faiss-cpu
